[Add] class付与 カスタムフック化
